Also support uservalidationbyemail
When uservalidationbyemail is also used -which can happen as there is no declared conflict-, admin validation does not validate the user from uservalidationbyemail point of view. 
It seems reasonable to assume that user should be validated once, and not wait for email validaiton if admin enables manually the account, so we need to set the metadata used by uservalidationbyemail : ->validated, and also ->validated_method so we keep track of validation method too.